backend/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.3.1
PyPDF2==3.0.1
google-generativeai==0.8.5
psycopg[binary]==3.2.3
python-dotenv==1.0.1
